decouple dialogues
