+++ configureHooks.go
@@ -1,12 +1,5 @@
 package akevitt
 
-// Called when engine.Dialogue is called
-func (builder *akevittBuilder) UseOnDialogue(f DialogueFunc) *akevittBuilder {
-	builder.engine.onDialogue = f
-
-	return builder
-}
-
 // Accepts function which gets called when the user lefts the game.
 // Note: use with caution, because calling methods from the engine like Message
 // will cause an infinite recursion
