+++ types.go
@@ -17,5 +17,3 @@ type UIFunc = func(engine *Akevitt, session *ActiveSession) tview.Primitive
 type CommandFunc = func(engine *Akevitt, session *ActiveSession, arguments string) error
 
 type DeadSessionFunc = func(deadSession *ActiveSession, liveSessions []*ActiveSession, engine *Akevitt)
-
-type DialogueFunc = func(engine *Akevitt, session *ActiveSession, dialogue *Dialogue) error
